Isolate map tables and resources per instance
