+++ fluent/messages.lua
@@ -1,8 +1,7 @@
 -- External dependencies
 local class = require("pl.class")
 
-local FluentMessages = class({
-  })
+local FluentMessages = class()
 
 return FluentMessages
